Unify local and remote notification services
