+++ test/model/notifications/fake_notification_service.dart
@@ -1,0 +1,25 @@
+import 'package:lichess_mobile/src/model/notifications/notification_service.dart';
+
+class FakeNotificationService implements NotificationService {
+  Map<int, LocalNotification> notifications = {};
+
+  @override
+  Future<void> start() async {}
+
+  @override
+  Future<void> registerDevice() async {}
+
+  @override
+  Future<void> unregister() async {}
+
+  @override
+  Future<void> cancel(int id) async {
+    notifications.remove(id);
+  }
+
+  @override
+  Future<int> show(LocalNotification notification) async {
+    notifications[notification.id] = notification;
+    return notification.id;
+  }
+}
